+++ src/.env.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+    const __KAPLAY_VERSION__: string;
+    const __KAPLAY_MAJORMINOR__: string;
+    const __KAPLAY_MAJOR__: string;
+    const __SITE_VERSION__: string;
+}