+++ CHANGELOG.md
@@ -6,7 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 11/3/2025
+## [1.1.1] - 2025-05-16
+
+### Added
+
+- Now you can see website version in the footer
+- Now the version in version selector is the same as the one in NPM
+
+### Fixed
+
+- Fixed a bug that made non current version guides being shown
+
+## [1.1.0] - 2025-3-11
 
 ### Added
 
@@ -14,7 +25,7 @@ and this project adheres to
 - Now in `/crew` you can select between outlined/original for downloading
 - Now imports appear `/crew` entries, as a new feature of crew package
 
-## [1.0.1] - 18/2/2025
+## [1.0.1] - 2025-2-18
 
 ### Added